requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.1
python-dotenv==1.0.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
cryptography==42.0.5
Pillow==10.2.0
numpy==1.26.4
imagehash==4.3.1
tqdm==4.66.2
dropbox==11.36.2
msal==1.26.0
requests==2.31.0
yfinance==0.2.36
gunicorn==21.2.0
aiohttp==3.9.1
tabulate==0.9.0
